requirements.txt
streamlit==1.31.0
pandas==2.1.0
numpy==1.26.0
plotly==5.18.0
fpdf==1.7.2
pillow==10.0.0
openpyxl==3.1.2
xlsxwriter==3.1.2
reportlab==4.0.4
python-docx==1.0.0
streamlit-option-menu==0.3.6
streamlit-calendar==0.1.0
streamlit-extras==0.3.0
